pyproject.toml
[tool.poetry]
name = "volunthub-backend"
version = "2.1.0"
description = "Backend de VoluntHub. Migracion a FastAPI con independencia de Supabase"
authors = ["Kevin <kevin@example.com>"]
readme = "README.md"
packages = [{ include = "app" }]

[tool.poetry.dependencies]
python = ">=3.11,<3.13"
fastapi = "^0.124.0"
uvicorn = { extras = ["standard"], version = "^0.29.0" }
strawberry-graphql = "^0.287.2"
asyncpg = "^0.31.0"
sqlalchemy = "^2.0.45"
alembic = "^1.17.2"
pydantic-settings = "^2.12.0"
python-dotenv = "^1.2.1"
geoalchemy2 = "^0.18.1"
shapely = "^2.1.2"
aiodataloader = "^0.4.3"

[tool.poetry.group.dev.dependencies]
pytest = "^8.0"
ruff = "^0.3"
black = "^24.0"
mypy = "^1.8"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
